TI restructured kernel build to be like other package builds in their repo, so shifting to this also.

+++ run.sh
@@ -26,10 +26,11 @@ source ${projdir}/version.sh
 
 mkdir -p ${builddir}
 
-if $custom_build ; then
-    run_custom_build
-    exit 0
-fi
+# RIAPS: Added to allow customization of the kernel config/dts
+#if $custom_build ; then
+#    run_custom_build
+#    exit 0
+#fi
 
 package_name=$(cd ${debcontroldir} && dpkg-parsechangelog --show-field Source)
 deb_version=$(cd ${debcontroldir} && dpkg-parsechangelog --show-field Version)
@@ -39,16 +40,41 @@ package_full="${package_name}-${package_version}"
 package_full_ll="${package_name}_${package_version}"
 echo "Building " $package_name " version " $deb_version
 
-if [ $require_root = "true" ] && [ "$EUID" -ne 0 ] ; then
-    echo "Requires root privileges to execute"
-    echo "Exiting"
-    exit 1
-fi
+# MM - for debugging
+echo "Package_full_ll: " $package_full_ll
+echo "Build dir: " $builddir
+echo "Source dir: " $sourcedir
+echo "Debcontroldir: " $debcontroldir
+echo "Package name: " $package_name
+echo "Deb Version: " $deb_version
+echo "Package Version: " $package_version
+echo "Last tested commit: " $last_tested_commit
 
-mkdir -p "${sourcedir}"
+# Setup Build Tools
+toolchain_dir = ${topdir}/tools/
+
+if [ -d "${topdir}/tools" ]; then
+    echo "Build tools already available in ${topdir}/tools"
+else
+    mkdir -p ${topdir}/tools/
+    cd ${topdir}/tools/
+    echo "> Downloading Aarch64 Toolchain .."
+    wget https://developer.arm.com/-/media/Files/downloads/gnu/13.2.rel1/binrel/arm-gnu-toolchain-13.2.rel1-x86_64-aarch64-none-linux-gnu.tar.xz &>>/dev/null
+    if [ $? -eq 0 ]; then
+        echo "> Aarch64 Toolchain: downloaded .."
+        tar -Jxf arm-gnu-toolchain-13.2.rel1-x86_64-aarch64-none-linux-gnu.tar.xz
+        rm arm-gnu-toolchain-13.2.rel1-x86_64-aarch64-none-linux-gnu.tar.xz
+        echo "> Aarch64 Toolchain: available"
+        export AARCH64_TOOL_LOC="$PWD/arm-gnu-toolchain-13.2.Rel1-x86_64-aarch64-none-linux-gnu/bin"
+        echo "$AARCH64_TOOL_LOC"
+    else
+        echo "> Aarch Toolchain: Failed to download. Exit code: $?"
+    fi
+fi
 
 # Generate original source tarball if none found
 if [ ! -f "${builddir}/${package_full_ll}.orig.tar.gz" ]; then
+    mkdir -p "${sourcedir}"
     if [ ! -d "${sourcedir}/${package_name}" ]; then
         git clone "${git_repo}" "${sourcedir}/${package_name}"
     fi
@@ -60,26 +86,44 @@ if [ ! -f "${builddir}/${package_full_ll}.orig.tar.gz" ]; then
       --transform "s,${sourcedir}/${package_name},${package_full},"
 fi
 
-# Extract original source tarball
-tar -xzmf "${builddir}/${package_full_ll}.orig.tar.gz" -C "${builddir}"
+# Generate source package if none found
+if [ ! -f "${builddir}/${package_name}_${deb_version}.dsc" ]; then
+    # Extract original source tarball
+    tar -xzmf "${builddir}/${package_full_ll}.orig.tar.gz" -C "${builddir}"
 
-# Add default debian control files if none found
-if [ ! -d "${builddir}/${package_full}/debian" ]; then
-    (cd "${builddir}/${package_full}" && debmake)
+    # Deploy our Debian control files
+    cp -rv "${debcontroldir}/debian" "${builddir}/${package_full}/"
+
+    # Build source package
+    (cd "${builddir}/${package_full}" && dpkg-source -b .)
+
+    # Cleanup intermediate source directory
+    #MM: keep files while debugging
+    #rm -r "${builddir}/${package_full}"
 fi
 
-# Deploy our Debian control files
-cp -rv "${debcontroldir}/debian" "${builddir}/${package_full}/"
+## Add default debian control files if none found
+#if [ ! -d "${builddir}/${package_full}/debian" ]; then
+#    (cd "${builddir}/${package_full}" && debmake)
+#fi
 
-run_prep || true
+# Generate binary package for this arch if not found
+build_arch=$(dpkg --print-architecture)
+if [ ! -f "${builddir}/${package_name}_${deb_version}_${build_arch}.buildinfo" ]; then
+    run_prep || true
 
-# Build source package
-(cd "${builddir}/${package_full}" && dpkg-source -b .)
+    # Extract source package
+    if [ ! -d "${builddir}/${package_name}_${deb_version}" ]; then
+        dpkg-source -x "${builddir}/${package_name}_${deb_version}.dsc" "${builddir}/${package_name}_${deb_version}"
+    fi
 
-# Extract source package
-if [ ! -d "${builddir}/${package_name}_${deb_version}" ]; then
-    dpkg-source -x "${builddir}/${package_name}_${deb_version}.dsc" "${builddir}/${package_name}_${deb_version}"
+    # Install build dependencies
+    (cd "${builddir}/${package_name}_${deb_version}" && mk-build-deps -ir -t "apt-get -o Debug::pkgProblemResolver=yes -y --no-install-recommends")
+
+    # Build binary package
+    (cd "${builddir}/${package_name}_${deb_version}" && debuild --no-lintian --no-sign || true)
+
+    # Cleanup intermediate build directory
+    #MM: keep files while debugging
+    #rm -r "${builddir}/${package_name}_${deb_version}"
 fi
-
-# Build binary package
-(cd "${builddir}/${package_name}_${deb_version}" && debuild --no-lintian --no-sign)
